+++ CHANGELOG.md
@@ -729,3 +729,4 @@
 * Change operators' precedence, length of operators define precedence
 * Remove `leaf`, `left`, `right`, `both` in `Binary` module
 * Define `Zipper` instance of `Nonempty Rose` tree
+* Change `Zipper` definition - add `structure` in tag